requirements_minimal.txt
composio
groq
pydantic
python-dotenv
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
PyMuPDF
pathlib